refactor(cli): remove legacy index.ts and implement cli.ts
